+++ internal/gui/gui.go
@@ -1,6 +1,3 @@
-//go:build systray
-// +build systray
-
 package gui
 
 import (
@@ -8,7 +5,6 @@ import (
 	"embed"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"net"
 	"net/http"
 	"strings"
@@ -70,7 +66,7 @@ func GetOnGUIReady(icon []byte, conf *visorconfig.V1) func() {
 	logger := logging.NewMasterLogger()
 	logger.SetLevel(logrus.InfoLevel)
 
-	httpC := getHTTPClient(conf, context.Background(), logger)
+	httpC := getHTTPClient(context.Background(), conf, logger)
 
 	if isRoot() {
 		return func() {
@@ -78,20 +74,19 @@ func GetOnGUIReady(icon []byte, conf *visorconfig.V1) func() {
 			systray.SetTooltip("Skywire")
 			initUIBtns(conf)
 			initVpnClientBtn(conf, httpC, logger)
-			initAdvancedButton(conf)
+			initAdvancedButton()
 			initQuitBtn()
-			go handleRootInteraction(conf, doneCh)
+			go handleRootInteraction(doneCh)
 		}
-	} else {
-		return func() {
-			systray.SetTemplateIcon(icon, icon)
-			systray.SetTooltip("Skywire")
-			initUIBtns(conf)
-			initVpnClientBtn(conf, httpC, logger)
-			initAdvancedButton(conf)
-			initQuitBtn()
-			go handleUserInteraction(conf, doneCh)
-		}
+	}
+	return func() {
+		systray.SetTemplateIcon(icon, icon)
+		systray.SetTooltip("Skywire")
+		initUIBtns(conf)
+		initVpnClientBtn(conf, httpC, logger)
+		initAdvancedButton()
+		initQuitBtn()
+		go handleUserInteraction(conf, doneCh)
 	}
 }
 
@@ -127,7 +122,7 @@ func Stop() {
 	systray.Quit()
 }
 
-func initAdvancedButton(conf *visorconfig.V1) {
+func initAdvancedButton() {
 	mAdvancedButton := systray.AddMenuItem("Advanced", "Advanced Menu")
 	mUninstall = mAdvancedButton.AddSubMenuItem("Uninstall", "Uninstall Application")
 
@@ -178,40 +173,29 @@ func initUIBtns(vc *visorconfig.V1) {
 }
 
 func initVpnClientBtn(conf *visorconfig.V1, httpClient *http.Client, logger *logging.MasterLogger) {
-	rpc_logger := logger.PackageLogger("systray:rpc_client")
-	rpcC = rpcClient(conf, rpc_logger)
+	rpcLogger := logger.PackageLogger("systray:rpc_client")
+	rpcC = rpcClient(conf, rpcLogger)
 
 	mVPNClient = systray.AddMenuItem("VPN", "VPN Client Submenu")
-	mVPNClient.Disable()
-	go mVPNClientEnable(rpcC, logger.PackageLogger("systray:vpn_button"))
 	// VPN Status
 	mVPNStatus = mVPNClient.AddSubMenuItem("Status: Disconnected", "VPN Client Status")
 	mVPNStatus.Disable()
-	go vpnStatusBtn(conf, rpcC)
+	go vpnStatusBtn(rpcC)
 	// VPN Connect/Disconnect Button
 	mVPNButton = mVPNClient.AddSubMenuItem("Connect", "VPN Client Switch Button")
 	// VPN Public Servers List
-	go serversBtn(conf, rpcC, httpClient, logger.PackageLogger("systray:vpn-servers"))
-}
-
-func mVPNClientEnable(rpcC visor.API, logger *logging.Logger) {
-	logger.Info("Initializing...")
-	for {
-		isReady, _ := rpcC.IsDMSGClientReady()
-		if isReady {
-			mVPNClient.Enable()
-			logger.Info("Initialized.")
-			return
-		}
-		logger.Warn("DMSG-Client not ready yet, will check in next 10 seconds.")
-		time.Sleep(10 * time.Second)
+	mVPNServersList := mVPNClient.AddSubMenuItem("Servers", "VPN Client Servers")
+	mVPNServers := []*systray.MenuItem{}
+	for _, server := range getAvailPublicVPNServers(conf, httpClient, logger.PackageLogger("systray:servers")) {
+		mVPNServers = append(mVPNServers, mVPNServersList.AddSubMenuItemCheckbox(server, "", false))
 	}
+	go serversBtn(mVPNServers, rpcC)
 }
 
-func vpnStatusBtn(conf *visorconfig.V1, rpcClient visor.API) {
+func vpnStatusBtn(rpcClient visor.API) {
 	for {
 		vpnStatusMx.Lock()
-		stats, _ := rpcClient.GetAppConnectionsSummary(skyenv.VPNClientName)
+		stats, _ := rpcClient.GetAppConnectionsSummary(skyenv.VPNClientName) //nolint
 		if len(stats) == 1 {
 			if stats[0].IsAlive {
 				if vpnLastStatus != 1 {
@@ -242,29 +226,11 @@ func vpnStatusBtn(conf *visorconfig.V1, rpcClient visor.API) {
 	}
 }
 
-func serversBtn(conf *visorconfig.V1, rpcClient visor.API, httpC *http.Client, logger *logging.Logger) {
-	// fetching and create vpn-servers lists
-	mVPNServersList := mVPNClient.AddSubMenuItem("Servers", "VPN Client Servers")
-	mVPNServers := []*systray.MenuItem{}
-	for {
-		servers, err := getAvailPublicVPNServers(conf, httpC, logger)
-		if err != nil {
-			logger.Warnf("cannot fetch vpn-servers due to: %s", err.Error())
-		}
-		if len(servers) > 0 {
-			for _, server := range servers {
-				mVPNServers = append(mVPNServers, mVPNServersList.AddSubMenuItemCheckbox(server, "", false))
-			}
-			break
-		}
-		time.Sleep(10 * time.Second)
-	}
-
-	// add btnChannel for each vpn server button
+func serversBtn(servers []*systray.MenuItem, rpcClient visor.API) {
 	btnChannel := make(chan int)
-	for index, server := range mVPNServers {
+	for index, server := range servers {
 		go func(chn chan int, server *systray.MenuItem, index int) {
-			for {
+			for { //nolint
 				select {
 				case <-server.ClickedCh:
 					chn <- index
@@ -273,12 +239,11 @@ func serversBtn(conf *visorconfig.V1, rpcClient visor.API, httpC *http.Client, l
 		}(btnChannel, server, index)
 	}
 
-	// logic of choosing server from list
 	for {
-		selectedServer := mVPNServers[<-btnChannel]
+		selectedServer := servers[<-btnChannel]
 		serverTempValue := strings.Split(selectedServer.String(), ",")[2]
 		serverPK := serverTempValue[2 : len(serverTempValue)-7]
-		for _, server := range mVPNServers {
+		for _, server := range servers {
 			server.Uncheck()
 			server.Enable()
 		}
@@ -289,24 +254,24 @@ func serversBtn(conf *visorconfig.V1, rpcClient visor.API, httpC *http.Client, l
 			continue
 		}
 
-		rpcClient.StopApp(skyenv.VPNClientName)
-		rpcClient.SetAppPK(skyenv.VPNClientName, pk)
+		rpcClient.StopApp(skyenv.VPNClientName)      //nolint
+		rpcClient.SetAppPK(skyenv.VPNClientName, pk) //nolint
 		vpnStatusMx.Lock()
 		vpnLastStatus = 3
 		vpnStatusMx.Unlock()
-		rpcClient.StartApp(skyenv.VPNClientName)
+		rpcClient.StartApp(skyenv.VPNClientName) //nolint
 	}
 }
 
-func handleVPNButton(conf *visorconfig.V1, rpcClient visor.API) {
-	stats, _ := rpcClient.GetAppConnectionsSummary(skyenv.VPNClientName)
+func handleVPNButton(rpcClient visor.API) {
+	stats, _ := rpcClient.GetAppConnectionsSummary(skyenv.VPNClientName) //nolint
 	if len(stats) == 1 {
-		rpcClient.StopApp(skyenv.VPNClientName)
+		rpcClient.StopApp(skyenv.VPNClientName) //nolint
 	} else {
 		vpnStatusMx.Lock()
 		vpnLastStatus = 3
 		vpnStatusMx.Unlock()
-		rpcClient.StartApp(skyenv.VPNClientName)
+		rpcClient.StartApp(skyenv.VPNClientName) //nolint
 	}
 }
 
@@ -325,7 +290,7 @@ func handleVPNLinkButton(conf *visorconfig.V1) {
 }
 
 // getAvailPublicVPNServers gets all available public VPN server from service discovery URL
-func getAvailPublicVPNServers(conf *visorconfig.V1, httpC *http.Client, logger *logging.Logger) ([]string, error) {
+func getAvailPublicVPNServers(conf *visorconfig.V1, httpC *http.Client, logger *logging.Logger) []string {
 	svrConfig := servicedisc.Config{
 		Type:     servicedisc.ServiceTypeVPN,
 		PK:       conf.PK,
@@ -336,7 +301,7 @@ func getAvailPublicVPNServers(conf *visorconfig.V1, httpC *http.Client, logger *
 	vpnServers, err := sdClient.Services(context.Background(), 0, "", "")
 	if err != nil {
 		logger.Error("Error getting vpn servers: ", err)
-		return nil, err
+		return nil
 	}
 	serverAddrs := make([]string, len(vpnServers))
 	for idx, server := range vpnServers {
@@ -346,12 +311,12 @@ func getAvailPublicVPNServers(conf *visorconfig.V1, httpC *http.Client, logger *
 			serverAddrs[idx] = server.Addr.PubKey().String() + " | NA"
 		}
 	}
-	return serverAddrs, nil
+	return serverAddrs
 }
 
-func getHTTPClient(conf *visorconfig.V1, ctx context.Context, logger *logging.MasterLogger) *http.Client {
+func getHTTPClient(ctx context.Context, conf *visorconfig.V1, logger *logging.MasterLogger) *http.Client {
 	var serviceURL dmsgget.URL
-	serviceURL.Fill(conf.Launcher.ServiceDisc)
+	serviceURL.Fill(conf.Launcher.ServiceDisc) //nolint
 	if serviceURL.Scheme == "dmsg" {
 		var keys cipher.PubKeys
 		servers := conf.Dmsg.Servers
@@ -365,7 +330,7 @@ func getHTTPClient(conf *visorconfig.V1, ctx context.Context, logger *logging.Ma
 		keys = append(keys, pk)
 		entries := direct.GetAllEntries(keys, servers)
 		dClient := direct.NewClient(entries, logger.PackageLogger("systray:dmsghttp_direct_client"))
-		dmsgDC, closeDmsg, err := direct.StartDmsg(ctx, logger.PackageLogger("systray:dmsghttp_dmsgDC"),
+		dmsgDC, closeDmsg, err := direct.StartDmsg(ctx, logger.PackageLogger("systray:dsmghttp_dmsgDC"),
 			pk, sk, dClient, dmsg.DefaultConfig())
 		if err != nil {
 			return &http.Client{}
@@ -401,19 +366,6 @@ func getHTTPClient(conf *visorconfig.V1, ctx context.Context, logger *logging.Ma
 	return &http.Client{}
 }
 
-func isSetVPNClientPKExist(conf *visorconfig.V1) bool {
-	for _, v := range conf.Launcher.Apps {
-		if v.Name == skyenv.VPNClientName {
-			for index := range v.Args {
-				if v.Args[index] == "-srv" {
-					return true
-				}
-			}
-		}
-	}
-	return false
-}
-
 func initQuitBtn() {
 	mQuit = systray.AddMenuItem("Quit", "")
 }
@@ -424,7 +376,7 @@ func handleUserInteraction(conf *visorconfig.V1, doneCh chan<- bool) {
 		case <-mOpenHypervisor.ClickedCh:
 			handleOpenHypervisor(conf)
 		case <-mVPNButton.ClickedCh:
-			handleVPNButton(conf, rpcC)
+			handleVPNButton(rpcC)
 		case <-mVPNLink.ClickedCh:
 			handleVPNLinkButton(conf)
 		case <-mUninstall.ClickedCh:
@@ -436,11 +388,11 @@ func handleUserInteraction(conf *visorconfig.V1, doneCh chan<- bool) {
 	}
 }
 
-func handleRootInteraction(conf *visorconfig.V1, doneCh chan<- bool) {
+func handleRootInteraction(doneCh chan<- bool) {
 	for {
 		select {
 		case <-mVPNButton.ClickedCh:
-			handleVPNButton(conf, rpcC)
+			handleVPNButton(rpcC)
 		case <-mUninstall.ClickedCh:
 			handleUninstall()
 		case <-mQuit.ClickedCh:
@@ -491,7 +443,7 @@ func stopVisor() {
 
 func isHypervisorRunning(addr string) bool {
 	// we check if it's up by querying `health` endpoint
-	resp, err := http.Get(addr)
+	resp, err := http.Get(addr) //nolint
 	if err != nil {
 		// hypervisor is not running in this case
 		return false
@@ -502,7 +454,7 @@ func isHypervisorRunning(addr string) bool {
 		}
 	}()
 
-	if _, err := io.Copy(ioutil.Discard, resp.Body); err != nil {
+	if _, err := io.Copy(io.Discard, resp.Body); err != nil {
 		log.WithError(err).Errorln("Failed to discard hypervisor response body")
 	}