+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 1.3.0
+
+### Added
+- add `--systray` flag to run visor as systray
+
+### Changed
+- remove `make build-systray` and `make build-windows-systray`, and just keep `make build` and `make build-windows`
+
 ## 1.2.0
 
 ### Added